Verbesserung der Preformenc

+++ src/Lib/SwingScreen.java
@@ -65,7 +65,7 @@ public class SwingScreen extends JPanel {
 
 
     @Override public Dimension getPreferredSize() {
-        return new Dimension((int)logic.width(),(int)logic.height());
+        return new Dimension(logic.width(), logic.height());
     }
